R6 virtual class for SBM representation

R6 virtual class for SBM representation

R6 virtual class for SBM representation

Super class

sbm::SBM -> BipartiteSBM

Active bindings

varProb

variational probabilities of nodes being in a block. Either return the probabilities of clustering or may be used to set that probabilities

memberships

list of size 2: vector of memberships in row, in column.

Methods

Public methods

BipartiteSBM$new()

BipartiteSBM$predict()

BipartiteSBM$reorder()

BipartiteSBM$clone()

Inherited methods

sbm::SBM$plot()

sbm::SBM$print()

sbm::SBM$show()

Method new()

constructor for a Simple SBM fit

Usage

BipartiteSBM$new(
  incidenceMatrix,
  model,
  dimLabels = list(row = "rowLabel", col = "colLabel"),
  covarList = list()
)

Arguments

incidenceMatrix

incidence (weighted) matrix

model

character ('bernoulli', 'poisson', 'gaussian')

dimLabels

list of labels of each dimension (in row, in columns)

covarList

and optional list of covariates, each of whom must have the same dimension as incidenceMatrix

directed

logical, directed network or not. In not, incidenceMatrix must be symmetric.

Method predict()

prediction under the current parameters

Usage

BipartiteSBM$predict(covarList = self$covarList)

Arguments

covarList

a list of covariates. By default, we use the covariates with which the model was estimated.

Method reorder()

permute group labels by order of decreasing probability

Usage

BipartiteSBM$reorder()

Method clone()

The objects of this class are cloneable with this method.

Usage

BipartiteSBM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.