R6 Class definition of an Bipartite SBM fit

R6 Class definition of an Bipartite SBM fit

R6 Class definition of an Bipartite SBM fit

Details

This class is designed to give a representation and adjust an LBM fitted with blockmodels.

The list of parameters estimOptions essentially tunes the optimization process and the variational EM algorithm, with the following parameters

"nbCores"integer for number of cores used. Default is 2

"verbosity"integer for verbosity (0, 1). Default is 1

"plot"boolean, should the ICL by dynamically plotted or not. Default is TRUE

"exploreFactor"control the exploration of the number of groups

"exploreMin"explore at least until exploreMin even if the exploration factor rule is achieved. Default 4. See the package blockmodels for details.

"exploreMin"Stop exploration at exploreMax even if the exploration factor rule is not achieved. Default Inf. See the package blockmodels for details.

"fast"logical: should approximation be used for Bernoulli model with covariates. Default to TRUE

Super classes

sbm::SBM -> sbm::BipartiteSBM -> BipartiteSBM_fit

Active bindings

loglik

double: approximation of the log-likelihood (variational lower bound) reached

ICL

double: value of the integrated classification log-likelihood

penalty

double, value of the penalty term in ICL

entropy

double, value of the entropy due to the clustering distribution

storedModels

data.frame of all models fitted (and stored) during the optimization

Methods

Public methods

BipartiteSBM_fit$new()

BipartiteSBM_fit$optimize()

BipartiteSBM_fit$setModel()

BipartiteSBM_fit$reorder()

BipartiteSBM_fit$show()

BipartiteSBM_fit$clone()

Inherited methods

sbm::SBM$print()

sbm::SBM$rNetwork()

sbm::BipartiteSBM$plot()

sbm::BipartiteSBM$predict()

sbm::BipartiteSBM$rEdges()

sbm::BipartiteSBM$rMemberships()

Method new()

constructor for a Bipartite SBM fit

Usage

BipartiteSBM_fit$new(
  incidenceMatrix,
  model,
  dimLabels = c(row = "rowName", col = "colName"),
  covarList = list()
)

Arguments

incidenceMatrix

rectangular (weighted) matrix

model

character ('bernoulli', 'poisson', 'gaussian')

dimLabels

labels of each dimension (in row, in columns)

covarList

and optional list of covariates, each of whom must have the same dimension as incidenceMatrix

Method optimize()

function to perform optimization

Usage

BipartiteSBM_fit$optimize(estimOptions = list())

Arguments

estimOptions

a list of parameters controlling the inference algorithm and model selection. See details.

Method setModel()

method to select a specific model among the ones fitted during the optimization. Fields of the current SBM_fit will be updated accordingly.

Usage

BipartiteSBM_fit$setModel(index)

Arguments

index

integer, the index of the model to be selected (row number in storedModels)

Method reorder()

permute group labels by order of decreasing probability

Usage

BipartiteSBM_fit$reorder()

Method show()

show method

Usage

BipartiteSBM_fit$show(type = "Fit of a Bipartite Stochastic Block Model")

Arguments

type

character used to specify the type of SBM

Method clone()

The objects of this class are cloneable with this method.

Usage

BipartiteSBM_fit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.