R6 class for Bipartite SBM sampler

R6 class for Bipartite SBM sampler

R6 class for Bipartite SBM sampler

Super classes

sbm::SBM -> sbm::SBM_sampler -> BipartiteSBM_sampler

Active bindings

nbNodes

vector of size 2: number of nodes (rows, columns)

nbBlocks

vector of size 2: number of blocks (rows, columns)

nbDyads

number of dyads (potential edges in the network)

memberships

list of size 2: vector of memberships in row, in column.

indMemberships

list of 2 matrix for clustering memberships

expectation

expected values of connection under the current model

Methods

Public methods

BipartiteSBM_sampler$new()

BipartiteSBM_sampler$rMemberships()

BipartiteSBM_sampler$rIncidence()

BipartiteSBM_sampler$show()

BipartiteSBM_sampler$clone()

Inherited methods

sbm::SBM$plot()

sbm::SBM$print()

Method new()

constructor for SBM

Usage

BipartiteSBM_sampler$new(
  model,
  nbNodes,
  blockProp,
  connectParam,
  dimLabels = list(row = "rowLabel", col = "colLabel"),
  covarParam = numeric(0),
  covarList = list()
)

Arguments

model

character describing the type of model

nbNodes

number of nodes in the network

blockProp

parameters for block proportions (vector of list of vectors)

connectParam

list of parameters for connectivity with a matrix of means 'mean' and an optional scalar for the variance 'var'. The dimensions of mu must match blockProp lengths

dimLabels

optional labels of each dimension (in row, in column)

covarParam

optional vector of covariates effect

covarList

optional list of covariates data

Method rMemberships()

a method to generate a vector of block indicators

Usage

BipartiteSBM_sampler$rMemberships()

Method rIncidence()

a method to sample an adjacency matrix for the current SBM

Usage

BipartiteSBM_sampler$rIncidence()

Returns

nothing (sampled matrix is store in the current object, accessible via $netMatrix)

Method show()

show method

Usage

BipartiteSBM_sampler$show(
  type = "Sampler for a Bipartite Stochastic Block Model"
)

Arguments

type

character used to specify the type of SBM

Method clone()

The objects of this class are cloneable with this method.

Usage

BipartiteSBM_sampler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.