R6 Class definition of a Multipartite SBM fit

R6 Class definition of a Multipartite SBM fit

R6 Class definition of a Multipartite SBM fit

Details

This class is designed to give a representation and adjust a Multipartite SBM fitted with GREMLIN.

Super classes

sbm::SBM -> sbm::MultipartiteSBM -> MultipartiteSBM_fit

Active bindings

loglik

double: approximation of the log-likelihood (variational lower bound) reached

ICL

double: value of the integrated classification log-likelihood

storedModels

data.frame of all models fitted (and stored) during the optimization

Methods

Public methods

MultipartiteSBM_fit$new()

MultipartiteSBM_fit$optimize()

MultipartiteSBM_fit$predict()

MultipartiteSBM_fit$setModel()

MultipartiteSBM_fit$show()

MultipartiteSBM_fit$clone()

Inherited methods

sbm::SBM$rNetwork()

sbm::MultipartiteSBM$plot()

sbm::MultipartiteSBM$print()

Method new()

constructor for Multipartite SBM

Usage

MultipartiteSBM_fit$new(netList)

Arguments

netList

list of SBM objects

Method optimize()

estimation of multipartiteSBM via GREMLINS

Usage

MultipartiteSBM_fit$optimize(estimOptions)

Arguments

estimOptions

options for MultipartiteBM

Method predict()

prediction under the currently estimated model

Usage

MultipartiteSBM_fit$predict()

Returns

a list of matrices matrix of expected values for each dyad

Method setModel()

method to select a specific model among the ones fitted during the optimization. Fields of the current MultipartiteSBM_fit will be updated accordingly.

Usage

MultipartiteSBM_fit$setModel(index)

Arguments

index

integer, the index of the model to be selected (row number in storedModels)

Method show()

show method

Usage

MultipartiteSBM_fit$show(type = "Fit of a Multipartite Stochastic Block Model")

Arguments

type

character used to specify the type of SBM

Method clone()

The objects of this class are cloneable with this method.

Usage

MultipartiteSBM_fit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.