R6 virtual class for SBM representation (mother class of Simple and Bipartite SBM fit and sampler)

R6 virtual class for SBM representation (mother class of Simple and Bipartite SBM fit and sampler)

R6 virtual class for SBM representation (mother class of Simple and Bipartite SBM fit and sampler)

Active bindings

dimension

size-2 vector: dimension of the network

modelName

character, the family of model for the distribution of the edges

dimLabels

vector or list of characters, the label of each dimension

nbCovariates

integer, the number of covariates

blockProp

block proportions (aka prior probabilities of each block)

connectParam

parameters associated to the connectivity of the SBM, e.g. matrix of inter/inter block probabilities when model is Bernoulli

covarParam

vector of regression parameters associated with the covariates.

covarList

list of matrices of covariates

covarArray

the array of covariates

covarEffect

effect of covariates

netMatrix

the matrix (adjacency or incidence) encoding the network

expectation

expected values of connection under the currently adjusted model

Methods

Public methods

SBM$new()

SBM$plot()

SBM$show()

SBM$print()

SBM$clone()

Method new()

constructor for SBM

Usage

SBM$new(
  model = "",
  dimension = numeric(2),
  dimLabels = list(row = NULL, col = NULL),
  blockProp = numeric(0),
  connectParam = list(mean = matrix()),
  covarParam = numeric(length(covarList)),
  covarList = list()
)

Arguments

model

character describing the type of model

dimension

dimension of the network matrix

dimLabels

labels of each dimension (in row, in columns)

blockProp

parameters for block proportions (vector of list of vectors)

connectParam

list of parameters for connectivity

covarParam

optional vector of covariates effect

covarList

optional list of covariates data

Method plot()

basic matrix plot method for SBM object or mesoscopic plot

Usage

SBM$plot(
  type = c("data", "expected", "meso"),
  ordered = TRUE,
  plotOptions = list()
)

Arguments

type

character for the type of plot: either 'data' (true connection), 'expected' (fitted connection) or 'meso' (mesoscopic view). Default to 'data'.

ordered

logical: should the rows and columns be reordered according to the clustering? Default to TRUE.

plotOptions

list with the parameters for the plot

Details

The list of parameters plotOptions for the mesoscopic plot is:

"seed": seed to control the layout

"title": character string for the title. Default value is NULL

"layout": Default value = NULL

"vertex.color": Default value is "salmon2"

"vertex.frame.color": Node border color.Default value is "black"

"vertex.shape": One of "none", "circle", "square", "csquare", "rectangle" "crectangle", "vrectangle", "pie", "raster", or "sphere". Default value = "circle"

"vertex.size": Size of the node (default is 2)

"vertex.size2": The second size of the node (e.g. for a rectangle)

"vertex.label.name": Names of the vertices. Default value is the label of the nodes

"vertex.label.color": Default value is "black"

"vertex.label.font": Default value is 2. Font: 1 plain, 2 bold, 3, italic, 4 bold italic, 5 symbol

"vertex.label.cex": Font size (multiplication factor, device-dependent).Default value is 0.9.

"vertex.label.dist": Distance between the label and the vertex. Default value is 0

"vertex.label.degree": The position of the label in relation to the vertex. default value is 0

"edge.threshold": Threshold under which the edge is not plotted. Default value is = -Inf

"edge.color": Default value is "gray"

"edge.width": Factor parameter. Default value is 10

"edge.arrow.size": Default value is 1

"edge.arrow.width": Default value is 2

"edge.lty": Line type, could be 0 or "blank", 1 or "solid", 2 or "dashed", 3 or "dotted", 4 or "dotdash", 5 or "longdash", 6 or "twodash". Default value is "solid"

"edge.curved": Default value is = 0.3.

For type = 'data' or 'expected plot', the list of parameters plotOptions is

"legend": Boolean. Set TRUE if you want to see the legend. Default value is FALSE

"legend.title":Boolean. Set TRUE if you want to print the title of the legend. Default value is FALSE

"legend.position":Position of the legend. Possible values are 'bottom', 'top','left,'right'. Default value is 'bottom'

"rowNames":Set true if the rownames must be plotted. Default value is FALSE

"colNames":Set true if the colNames must be plotted. Default value is FALSE

"line.color": Chain of character. The color of the lines to separate groups if a clustering is provided. Default value is red

"line.width": Numeric. Width of the lines to separate groups. Default value is NULL, automatically chosen

"title": Chain of character. Title of the plot. Default value is NULL

Returns

a ggplot2 object for the 'data' and 'expected', a list with the igraph object g, the layout and the plotOptions for the 'meso'

Method show()

print method

Usage

SBM$show(type = "Stochastic Block Model")

Arguments

type

character to tune the displayed name

Method print()

print method

Usage

SBM$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

SBM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.