R6 virtual class for SBM fit (mother class of Simple and Bipartite SBM fit)

R6 virtual class for SBM fit (mother class of Simple and Bipartite SBM fit)

R6 virtual class for SBM fit (mother class of Simple and Bipartite SBM fit)

Super class

sbm::SBM -> SBM_fit

Active bindings

probMemberships

matrix -- or list of 2 matrices for Bipartite network -- of estimated probabilities for block memberships for all nodes

loglik

double: approximation of the log-likelihood (variational lower bound) reached

ICL

double: value of the integrated classification log-likelihood

fitted

matrix of predicted value of the network

Methods

Public methods

SBM_fit$new()

SBM_fit$show()

SBM_fit$setModel()

SBM_fit$predict()

SBM_fit$reorder()

SBM_fit$clone()

Inherited methods

sbm::SBM$plot()

sbm::SBM$print()

Method new()

constructor for SBM fit

Usage

SBM_fit$new(data, model, dimLabels, covarList)

Arguments

data

the data matrix of the network

model

character describing the type of model

dimLabels

labels of each dimension (in row, in columns)

covarList

optional list of matrices for covariates

Method show()

print/show method

Usage

SBM_fit$show(type = "Fit of a Stochastic Block Model")

Arguments

type

character to tune the displayed name

Method setModel()

method to select a specific model among the ones fitted during the optimization. Fields of the current SBM_fit will be updated accordingly.

Usage

SBM_fit$setModel(index)

Arguments

index

integer, the index of the model to be selected (row number in storedModels)

Method predict()

prediction under the currently parameters

Usage

SBM_fit$predict(covarList = self$covarList)

Arguments

covarList

a list of covariates. By default, we use the covariates with which the model was estimated

Returns

a matrix of expected values for each dyad

Method reorder()

permute group labels by order of decreasing probability

Usage

SBM_fit$reorder()

Method clone()

The objects of this class are cloneable with this method.

Usage

SBM_fit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.