R6 Class definition of an Simple SBM fit

R6 Class definition of an Simple SBM fit

R6 Class definition of an Simple SBM fit

Details

This class is designed to give a representation and adjust an SBM fitted with blockmodels.

Super classes

sbm::SBM -> sbm::SBM_fit -> SimpleSBM_fit

Active bindings

nbNodes

number of nodes

nbBlocks

number of blocks

nbDyads

number of dyads (potential edges in the network)

memberships

vector of clustering

directed

is the network directed or not

storedModels

data.frame of all models fitted (and stored) during the optimization

Methods

Public methods

SimpleSBM_fit$new()

SimpleSBM_fit$optimize()

SimpleSBM_fit$predict()

SimpleSBM_fit$show()

SimpleSBM_fit$clone()

Inherited methods

sbm::SBM$plot()

sbm::SBM$print()

sbm::SBM_fit$setModel()

Method new()

constructor for a Simple SBM fit

Usage

SimpleSBM_fit$new(adjacencyMatrix, model, directed, covarList = list())

Arguments

adjacencyMatrix

square (weighted) matrix

model

character ('bernoulli', 'poisson', 'gaussian')

directed

logical, directed network or not. In not, adjacencyMatrix must be symmetric.

covarList

and optional list of covariates, each of whom must have the same dimension as adjacencyMatrix

Method optimize()

function to perform optimization

Usage

SimpleSBM_fit$optimize(
  verbosity = 3,
  plot = TRUE,
  explorFactor = 1.5,
  nbBlocksRange = c(4, Inf),
  nbCores = parallel::detectCores()
)

Arguments

verbosity

integer, the level of verbosity. Default to 3

plot

logical, if TRUE ploting is done dynamically on the screen. Default to TRUE

explorFactor

double factor for exploring successive model

nbBlocksRange

2-size vector: range of exploration

nbCores

integer, the number of cores to use. Default is parallel::detectCores().

Method predict()

prediction under the currently estimated model

Usage

SimpleSBM_fit$predict(covarList = self$covarList)

Arguments

covarList

a list of covariates. By default, we use the covariates with which the model was estimated

Returns

a matrix of expected values for each dyad

Method show()

show method

Usage

SimpleSBM_fit$show(type = "Fit of a Simple Stochastic Block Model")

Arguments

type

character used to specify the type of SBM

Method clone()

The objects of this class are cloneable with this method.

Usage

SimpleSBM_fit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.