R6 Class definition of a Multiplex SBM fit

R6 Class definition of a Multiplex SBM fit

R6 Class definition of a Multiplex SBM fit

Details

This class is designed to give a representation and adjust a Multiplex SBM fitted with GREMLIN.

The list of parameters estimOptions essentially tunes the optimization process and the variational EM algorithm, with the following parameters

"nbCores" integer for number of cores used. Default is 2

"verbosity" integer for verbosity (0, 1). Default is 1

"nbBlocksRange" List of length the number of functional groups, each element supplying the minimal and maximal number of blocks to be explored. The names of the list must be the names of the functional groups. Default value is from 1 to 10)

"initBM" Boolean. True if using simple and bipartite SBM as initialisations. Default value = TRUE

"maxiterVEM" Number of max. number of iterations in the VEM. Default value = 100

"maxiterVE" Number of max. number of iterations in the VE. Default value = 100

Super classes

sbm::SBM -> sbm::MultipartiteSBM -> sbm::MultipartiteSBM_fit -> MultiplexSBM_fit

Active bindings

nbBlocks

vector of size 2: number of blocks (rows, columns)

dependentNetwork

: connection parameters in each network

storedModels

data.frame of all models fitted (and stored) during the optimization

namesLayers

: names of the various Networks

Methods

Public methods

MultiplexSBM_fit$new()

MultiplexSBM_fit$optimize()

MultiplexSBM_fit$plot()

MultiplexSBM_fit$show()

MultiplexSBM_fit$predict()

MultiplexSBM_fit$clone()

Inherited methods

sbm::SBM$rNetwork()

sbm::MultipartiteSBM$print()

sbm::MultipartiteSBM_fit$setModel()

Method new()

constructor for Multiplex SBM

Usage

MultiplexSBM_fit$new(netList, dependentNet = FALSE)

Arguments

netList

list of SBM object with

dependentNet

boolean indicating whether dependence is assumed between networks beyond the common dependence on the latent variables

Method optimize()

estimation of multipartiteSBM via GREMLINS

Usage

MultiplexSBM_fit$optimize(estimOptions)

Arguments

estimOptions

options for MultipartiteBM

Method plot()

plot Multiplex Network

Usage

MultiplexSBM_fit$plot(
  type = c("data", "expected"),
  ordered = TRUE,
  plotOptions = list()
)

Arguments

type

character for the type of plot: either 'data' (true connection), 'expected' (fitted connection). Default to 'data'.

ordered

TRUE is the matrices are plotted after reorganization with the blocks. Default value = TRUE

plotOptions

list of plot options for the matrix view

Method show()

show method

Usage

MultiplexSBM_fit$show(type = "Fit of a Multiplex Stochastic Block Model")

Arguments

type

character used to specify the type of SBM

Method predict()

prediction under the currently estimated model

Usage

MultiplexSBM_fit$predict()

Returns

a list of matrices matrix of expected values for each dyad

Method clone()

The objects of this class are cloneable with this method.

Usage

MultiplexSBM_fit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.