R6 virtual class for SBM representation (mother class of Simple and Bipartite SBM fit and sampler)

R6 virtual class for SBM representation (mother class of Simple and Bipartite SBM fit and sampler)

R6 virtual class for SBM representation (mother class of Simple and Bipartite SBM fit and sampler)

Active bindings

dimension

size-2 vector: dimension of the network

modelName

character, the family of model for the distribution of the edges

dimLabels

vector of characters, the label of each dimension

nbCovariates

integer, the number of covariates

blockProp

vector of block proportions (aka prior probabilities of each block)

connectParam

parameters associated to the connectivity of the SBM, e.g. matrix of inter/inter block probabilities when model is Bernoulli

covarParam

vector of regression parameters associated with the covariates.

covarList

list of matrices of covariates

covarEffect

effect of covariates

netMatrix

the matrix (adjacency or incidence) encoding the network

Methods

Public methods

SBM$new()

SBM$plot()

SBM$show()

SBM$print()

SBM$clone()

Method new()

constructor for SBM

Usage

SBM$new(
  model = "",
  dimension = numeric(2),
  dimLabels = vector("list", 2),
  blockProp = numeric(0),
  connectParam = list(mean = matrix()),
  covarParam = numeric(length(covarList)),
  covarList = list()
)

Arguments

model

character describing the type of model

dimension

dimension of the network matrix

dimLabels

labels of each dimension (in row, in columns)

blockProp

parameters for block proportions (vector of list of vectors)

connectParam

list of parameters for connectivity

covarParam

optional vector of covariates effect

covarList

optional list of covariates data

Method plot()

basic matrix plot method for SBM object

Usage

SBM$plot(type = c("data", "expected"), ordered = TRUE)

Arguments

type

character for the type of plot: either 'data' (true connection) or 'expected' (fitted connection). Default to 'data'.

ordered

logical: should the rows and columns be reordered according to the clustering? Default to TRUE.

Returns

a ggplot2 object

Method show()

print method

Usage

SBM$show(type = "Stochastic Block Model")

Arguments

type

character to tune the displayed name

Method print()

print method

Usage

SBM$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

SBM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.